+++ tests/test_views_permissions.py
@@ -12,8 +12,16 @@ RECORD_ID = 'd5db6e57-2c10-43e2-96c8-56602ef01435'
 
 class PermissionsViewTest(BaseWebTest, unittest.TestCase):
 
+    def get_app_settings(self, extras=None):
+        settings = super(PermissionsViewTest, self).get_app_settings(extras)
+        settings['experimental_permissions_endpoint'] = 'True'
+        return settings
+
+
+class EntriesTest(PermissionsViewTest):
+
     def setUp(self):
-        super(PermissionsViewTest, self).setUp()
+        super(EntriesTest, self).setUp()
         self.app.put_json('/buckets/beers', MINIMALIST_BUCKET,
                           headers=self.headers)
         self.app.put_json('/buckets/beers/collections/barley',
@@ -29,11 +37,6 @@ class PermissionsViewTest(BaseWebTest, unittest.TestCase):
         # Other user.
         self.app.put_json('/buckets/water', MINIMALIST_BUCKET,
                           headers=get_user_headers('alice'))
-
-    def get_app_settings(self, extras=None):
-        settings = super(PermissionsViewTest, self).get_app_settings(extras)
-        settings['experimental_permissions_endpoint'] = 'True'
-        return settings
 
     def test_permissions_list_entries_for_current_principal(self):
         resp = self.app.get('/permissions', headers=self.headers)
@@ -91,3 +94,84 @@ class PermissionsViewTest(BaseWebTest, unittest.TestCase):
         self.assertEqual(resp.headers['Total-Records'], '4')
         self.assertIn('Next-Page', resp.headers)
         self.assertEqual(len(resp.json['data']), 2)
+
+
+class GroupsPermissionTest(PermissionsViewTest):
+
+    def setUp(self):
+        super(GroupsPermissionTest, self).setUp()
+
+        self.admin_headers = get_user_headers('admin')
+        self.admin_principal = self.app.get('/', headers=self.admin_headers).json['user']['id']
+
+        self.app.put_json('/buckets/beers',
+                          {'permissions': {'write': ['/buckets/beers/groups/admins']}},
+                          headers=self.headers)
+        self.app.put_json('/buckets/beers/groups/admins',
+                          {'data': {'members': [self.admin_principal]}},
+                          headers=self.headers)
+        self.app.put_json('/buckets/beers/collections/barley',
+                          MINIMALIST_COLLECTION,
+                          headers=self.headers)
+
+    def test_permissions_granted_via_groups_are_listed(self):
+        resp = self.app.get('/permissions', headers=self.admin_headers)
+        buckets = [e for e in resp.json['data'] if e['resource_name'] == 'bucket']
+        self.assertEqual(buckets[0]['id'], 'beers')
+        self.assertIn('write', buckets[0]['permissions'])
+
+    def test_permissions_inherited_are_not_listed(self):
+        resp = self.app.get('/permissions', headers=self.admin_headers)
+        collections = [e for e in resp.json['data'] if e['resource_name'] == 'collection']
+        self.assertEqual(len(collections), 0)
+
+
+class SettingsPermissionsTest(PermissionsViewTest):
+
+    admin_headers = get_user_headers('admin')
+    admin_principal = 'basicauth:bb7fe7b98e759578ef0de85b546dd57d21fe1e399390ad8dafc9886043a00e5c'  # NOQA
+
+    def __init__(self, *args, **kwargs):
+        super(SettingsPermissionsTest, self).__init__(*args, **kwargs)
+
+    def get_app_settings(self, extras=None):
+        settings = super(SettingsPermissionsTest, self).get_app_settings(extras)
+        settings['bucket_write_principals'] = 'system.Authenticated'
+        settings['group_create_principals'] = self.admin_principal
+        settings['collection_write_principals'] = 'system.Authenticated'
+        settings['record_create_principals'] = '/buckets/beers/groups/admins'
+        return settings
+
+    def setUp(self):
+        super(SettingsPermissionsTest, self).setUp()
+        self.app.put_json('/buckets/beers', MINIMALIST_BUCKET, headers=self.headers)
+        self.app.put_json('/buckets/beers/groups/admins',
+                          {'data': {'members': [self.admin_principal]}},
+                          headers=self.headers)
+        self.app.put_json('/buckets/beers/collections/barley',
+                          MINIMALIST_COLLECTION,
+                          headers=self.headers)
+
+    def test_bucket_write_taken_into_account(self):
+        resp = self.app.get('/permissions', headers=get_user_headers("any"))
+        buckets = [e for e in resp.json['data'] if e['resource_name'] == 'bucket']
+        self.assertEqual(buckets[0]['id'], 'beers')
+        self.assertIn('write', buckets[0]['permissions'])
+
+    def test_collection_create_taken_into_account(self):
+        resp = self.app.get('/permissions', headers=self.admin_headers)
+        buckets = [e for e in resp.json['data'] if e['resource_name'] == 'bucket']
+        self.assertEqual(buckets[0]['id'], 'beers')
+        self.assertIn('group:create', buckets[0]['permissions'])
+
+    def test_collection_write_taken_into_account(self):
+        resp = self.app.get('/permissions', headers=get_user_headers("any"))
+        collections = [e for e in resp.json['data'] if e['resource_name'] == 'collection']
+        self.assertEqual(collections[0]['id'], 'barley')
+        self.assertIn('write', collections[0]['permissions'])
+
+    def test_record_create_taken_into_account(self):
+        resp = self.app.get('/permissions', headers=self.admin_headers)
+        collections = [e for e in resp.json['data'] if e['resource_name'] == 'collection']
+        self.assertEqual(collections[0]['id'], 'barley')
+        self.assertIn('record:create', collections[0]['permissions'])