+++ CHANGELOG.rst
@@ -6,7 +6,9 @@ This document describes changes between each past release.
 4.3.3 (unreleased)
 ------------------
 
-- Nothing changed yet.
+**Bug fixes**
+
+- Permissions endpoint now exposes the user permissions defined in settings (fixes #909)
 
 
 4.3.2 (2016-10-25)